move hexlify methods into a new file
