+++ tests/unit_tests/libKitsunemimiCommon/common_methods/string_methods_test.h
@@ -28,7 +28,6 @@ private:
     void ltrim_test();
     void rtrim_test();
     void trim_test();
-    void hexlify_test();
 };
 
 } // namespace Kitsunemimi
