+++ tests/unit_tests/main.cpp
@@ -15,6 +15,7 @@
 
 #include <libKitsunemimiCommon/common_methods/string_methods_test.h>
 #include <libKitsunemimiCommon/common_methods/vector_methods_test.h>
+#include <libKitsunemimiCommon/common_methods/object_methods_test.h>
 
 #include <libKitsunemimiCommon/state_test.h>
 #include <libKitsunemimiCommon/statemachine_test.h>
@@ -32,7 +33,8 @@ int main()
     Kitsunemimi::StackBuffer_Test();
 
     Kitsunemimi::StringMethods_Test();
-    Kitsunemimi::TextFile_Test();
+    Kitsunemimi::VectorMethods_Test();
+    Kitsunemimi::ObjectMethods_Test();
 
     Kitsunemimi::State_Test();
     Kitsunemimi::Statemachine_Test();
